Create SAX content handler for symlinks files
Update AutomaticSymlinkUtility to take an XML symlinks definition file
as an argument and parse it using SAX to determine the links to be
created. Add a createSymbolicLink() method which will take the link and
target paths needed to create the symbolic link, this method will be
implemented later so will only log the link to be created.
Update AutomaticSymlinkUtilityTest to add tests for these changes.

Create SymlinksHander as an extension of DefaultHandler, this will
implement the parsing for symlinks definition files. The XML should be
processing and the method called to create the symbolic links.
Create SymlinksHandlerTest to add tests for SymlinksHandler.

Create SymlinkCreationResult to hold a result status and message from
the symbolic link creation.

Update build.gradle to include CheckStyle, FindBugs and PMD code quality
tools as well as JMockit for mocking and code coverage metrics.
Create configurations for CheckStyle and FindBugs, the FindBugs
configuration will filter some common false positives which occur with
JUnit and JMockit frameworks.