Implement link is existing directory scenarios
Update AutomaticSymlinkUtility to handle the link path is existing
directory scenarios for each type of target scenario.

Update AutomaticSymlinkUtilityTest to create tests for the link path is
existing directory scenarios.

Create MoveDirectoryVisitor to move a link folder to the target
location.

Create MoveDirectoryVisitorTest to create unit tests for the
MoveDirectoryVisitor class.